Reduce duplicates by standardizing dispositions (#1)
13343 duplicates removed (80574 to 67231).
From 780 distinct dispositions to 85 (7 of which are standards).

Introduced two standard codes in addition to the ones from Open Data
dataset, abandoned and transferred. These likely fall under excluded,
but seemed worth preserving. We can always make a "for analysis" set
that keeps those, while the main dataset (to be rid of duplicates)
standardizes them further.