Run deploy Action after update Action
Commits made by a GH Action (using the built-in GH creds) don't
trigger other Actions, to avoid recursive calls.[1] In this case,
though, we know we want the scheduled update to prompt a re-deploy,
since it'll most likely contain new entries (and if it doesn't, it's
not harmful to redeploy).

Didn't want to set up a PAT, since that seemed like more of a risk
than this constrained workflow call.

[1]: https://docs.github.com/en/actions/using-workflows/triggering-a-workflow#triggering-a-workflow-from-a-workflow